Change to use original Cover and Gallery images
These are a bit larger, sharper, not cropped and less wide so they look better in the scenes view. I also added info in comments for the real original images but those are huge and easily 5Mb+ in size. Probably too much right now but might be an option for later if XBVR could resize images to be permanently stored.