Fixed title, cover and sceneID for TmwVRnet scraper
This scraper had the same issue as RealJamVR, the deo player element was changed.
Also I'd like to note that on line 45, the split result tmp[5] will cause the app to crash if that doesn't exist. Which could easily happen if they change that element again, but I couldn't find a way to throw an error so a scraper is skipped on fatal errors like that.