VirtualRealPorn(/Trans) scraper: fix scraped filenames
VirtualRealPorn seems to employ a myriad of filename patterns, which to some extent have changed over time. Please let me know if I have forgotten a pattern.

Since I do not have an account and cannot verify the actual naming of the files, I made an educated guess for the filenames of VirtualRealTrans.