scrapers: Added more JAV scrapers (javland, javlibrary, javbus) (#1100)
* javr: enable multiple scraper options, and add javlibrary and javbus

* javr: add jav.land as well

* Update javdatabase.go

Add `javdatabase` as a tag

* Update javbus.go

Always add `javbus` as a tag

* Update javland.go

Always add 'jav.land' as a tag

* Update javlibrary.go

Always add javlibrary as a tag

* Update javtags.go

Changed both of the skip/re-map lists to tab separation.

Moved "solo/solo work/solowork" to the "drop" list, I forgot what the R18 tag was, but unless I am mistaken, it was a tag they (and FANZA by extension) automatically add(ed) to titles that aren't part of an overarching "series". It's meaningless to us since we don't scrape R18/DMM's "series" listings, nor would we have a way to filter for them in XBVR.

Changed "kiss kiss" to the tag to be retained, as that was the tag R18 used, and most users would already have plenty of in their databases. Probably best to maintain continuity with the old R18 tags whenever possible if this is to be done.

* Update javtags.go

Similarly, `suntan` was the tag R18 used, I have 12 entries in my library that pre-date the manifests I started writing myself.

* javr: more error-correcting code, and less code-reuse between scrapers

Co-authored-by: vt-idiot <81622808+vt-idiot@users.noreply.github.com>