scraper: Validate cover image for VirtualReal websites
Some of the covers provided by the VirtualReal websites are using a
pixel format that is not compatible with golang's `image/jpeg` decoder
and also with some VR players like HereSphere. It should be kinda safe
to assume that if a image can't be decoded by the golang decoders it may
not work in some players.

This patch validates the cover images against the golang decoders, and
tries to use some gallery image as cover, if no valid coner image is
found.

This is the only scraper I had this issue so far, but more scrapers may
need the same changes.