feat: add Has Image Filter
* Update RescrapeButton.vue

Updated the Rescrape Scene function to work with scenes that have scrape_id and belong to a configured scraper. By toggling it's needs_update and running that scraper.

It can be useful in case of single scenes without images.

I've had to add a small delay (200ms) as it was not triggering the forceUpdate when making the API call, even when checking if this.item.needs_update was true.

* Replaced github.com/cosmtrek/air with github.com/air-verse/air

The repo github.com/cosmtrek/air doesn't exist anymore. When running Gitpod, it suggests using github.com/air-verse/air instead.

* Replaced github.com/cosmtrek/air with github.com/air-verse/air

The repo github.com/cosmtrek/air doesn't exist anymore. When running Gitpod, it suggests using github.com/air-verse/air instead.

* Added the missing Has Image case