Ready for PR
Their some optimizations that could be done in the XPATH query. It usually can pull all scenes in the first scrape. Occasionally something will slow down the network and cause a misfire. And result in a few mixed scenes. Needs to be test on a slow network. Currently hangs if something really goes wrong and can't find wants it looking for