Should fix StashDB locked errors
Since the stashDB scraper is called directly via `scrape.StashDb()` through `api.StashDB()` and not though `tasks.Scrape()` All the locked code that is associated with `tasks.Scrape()` is not used. So DB acitions in `stashDB()` are being blocked. I believe this is causing errors when multiple process attempt to run as other processes lock the DB but stashDB does not and doesn't check either.

Also updated the start code to bring it more in line with how `tasks.Scrape()` starts as `scrapeID` and `siteID` are only used in this one instance.

Testing required to confirm squash