scraper: SinsVR cast fix (#1823)
* SinsVR Fix for model not present on scrape

Occasionally sinsVR will leave out the model on the scene page resulting in no model being listed for the scene. This falls back to using the trailer URL which contains the models in the scenes

* Debug prompt removal

* Some more tweaks. 

Check to ensure the profile url is valid. Notice occasionally there is a misspelling of the actor name in the url

* Format