Bug Fix & Code Optimatzions
Change from one run on function input to use a type struct. Makes the code cleaner and easer to read.

Another attempt at squashing the API returning results not belong to the correct studio. This should work has the code now checks to ensure the result belongs to the correct scraper. And ignores the rest.

Also changed the scraperID to a code used in the API result