Rebuild of Reality Lovers
API appears to have been abandoned and now returns 403 errors. This new scraper now relies on HTML for its data. All works. Title and Cover are both grabbed from index page. These seem to be the most accurate and useful data source. On single scene scrape we fall back to using the gallery and the URL for these sources