scraper: Add BrazzersVR Scraper (#1852)
* Add BrazzersVR Scraper

Appears to currently work. More wide spread testing needed. Had some randomness to what scenes were being shown on the index page. The current set of URLs appears to return the correct scenes.

The next index page does not appear to show up in the colly request only on a working browser. Had to resort to checking if any scene links where available and advancing if true.

* Go fmt & Cleanup

* BazzersVR v2

Abandoned the OG BazzersVR scraper in favor of the backend API call. Which shares the same API and JSON structure as VirtualPorn. All seems to be in working order.

Some of the code for VirtualPorn may still need to be stripped or adjusted based on differing sites. Current things that need be check or adjusted Filenames and Member Link

* Remove Debugging Prompt

* Go fmt fixes

* Code Optimizations & Bug Fix

The Origin and Refer are important. It should prevent the collisions in the API call when both scrapers are ran at the same time.  AKA VirtualPorn returning scene data for Brazzers

Also updated for parallel scraper optimization

* Bug Fix & Code Optimatzions

Change from one run on function input to use a type struct. Makes the code cleaner and easer to read.

Another attempt at squashing the API returning results not belong to the correct studio. This should work has the code now checks to ensure the result belongs to the correct scraper. And ignores the rest.

Also changed the scraperID to a code used in the API result

* Comment cleanup