scraper: Fix KinkVR scraper (#1875)
* Final KinkVR working

File names are dropped upon migration. Appears to have to good means for the future of guessing the filename. Do not need migration code. Due to the change in the URL all scenes are resraped on next update and all scraped data is replaced with its new format

* Bug fix

Bootstrap sites use a trailing slash. I remove this for internal logic and this is what is saved. But when comparing new scenes to what has been search the trailing slash caused this logic to never go true.

* Spin off kinkvr completely

After some digging it appears kinkvr was sold and not a website redesign by badoink. Now completly separates kinkvr from badoink

* Favicon Update

* Minor Code Clean up

Missed some remnants of badoinks colab with kink. Also removed scraper for actor has no info for actor is on kink just links to other scenes