fix: NAVR - real girls now
"Real Girls Now" uses a different element for trailer image used for covers. There are now 3 different formats. Reduced this code to a few "if" statements verses the duplicated filename generation code in existing foreach "loops". There should only be one possibility for this so no need to code "loops" for each format.